requirements.txt
# LLM & orchestration
anthropic>=0.42.0
langgraph>=0.3.0
langchain-core>=0.3.0

# Embeddings & vector store
openai>=1.50.0
chromadb>=0.5.0

# Interface
streamlit>=1.40.0

# Ingestion
requests>=2.31.0
beautifulsoup4>=4.12.0

# Evaluation
ragas>=0.2.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0

# Tokenization (chunker)
tiktoken>=0.7.0

# Config
python-dotenv>=1.0.0
